mybackend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
python-dotenv==1.0.0
